requirements.txt
Django>=5.0
psycopg2-binary
python-decouple
dj-database-url
django-allauth
whitenoise
django-crispy-forms
crispy-bootstrap5